requirements.txt
aiogram==3.3.0
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
apscheduler==3.10.4
requests==2.31.0
beautifulsoup4==4.12.2
python-dotenv==1.0.0
